refactor(plan-mode): split oversized test suite
