+++ extensions/pi-plan-mode/test/question-tool.test.ts
@@ -1,0 +1,58 @@
+import assert from "node:assert/strict";
+import test from "node:test";
+import { createMockContext, createMockPi } from "../../../test/support.js";
+import planMode, { normalizePlanModeQuestionParams } from "../src/plan-mode.js";
+
+test("plan_mode_question reports non-interactive cancellation", async () => {
+	const mock = createMockPi();
+	planMode(mock.pi);
+	const execute = mock.tools[0]?.execute as
+		| ((...args: unknown[]) => Promise<{ details?: { reason?: string } }>)
+		| undefined;
+	assert.ok(execute);
+	const context = createMockContext({ hasUI: false });
+	await mock.commands.get("plan")?.handler("", context.ctx);
+	const result = await execute(
+		"call-1",
+		{
+			questions: [
+				{
+					id: "scope",
+					header: "Scope",
+					question: "How broad?",
+					options: [
+						{ label: "Small", description: "Only the bug." },
+						{ label: "Broad", description: "Include cleanup." },
+					],
+				},
+			],
+		},
+		undefined,
+		undefined,
+		context.ctx,
+	);
+	assert.equal(result.details?.reason, "ui_unavailable");
+});
+
+test("normalizePlanModeQuestionParams validates question shape", () => {
+	const result = normalizePlanModeQuestionParams({
+		questions: [
+			{
+				id: "scope",
+				header: "Scope",
+				question: "How broad?",
+				options: [
+					{ label: "Small", description: "Only the bug." },
+					{ label: "Broad", description: "Include nearby cleanup." },
+				],
+			},
+		],
+	});
+
+	assert.equal(result.ok, true);
+	if (result.ok) assert.equal(result.questions[0]?.options[1]?.label, "Broad");
+	assert.deepEqual(normalizePlanModeQuestionParams({ questions: [] }), {
+		ok: false,
+		error: "questions must contain 1-3 items",
+	});
+});
